requirements.txt
boto3==1.28.1
botocore==1.31.1
cachetools==5.3.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.6
pyjwt==2.8.0

google-cloud-core==2.3.3
google-cloud-storage==2.10.0